feat: finished adding css to the login view. Made a css modules folder with a css file. Removed a lot of the extra css from app.css and index.css
