+++ src/app.js
@@ -2,22 +2,22 @@ import React from "react";
 import { Route, Switch } from "react-router-dom";
 import { useAuth0 } from "@auth0/auth0-react";
 import { NavBar, Footer } from "./components";
-import { Home } from "./views";
 import "./app.css";
 import LoginView from "./views/loginView";
+import Home from "./views/homeView";
 
 const App = () => {
   const { isAuthenticated } = useAuth0();
   
   if(isAuthenticated){
     return (
-      <div id="app">
+      <div class="mainApp">
             <Route path="/" exact component={Home} />
       </div>
     );
   }else{
     return (
-      <div id="app">
+      <div class="mainApp">
             <Route path="/" exact component={LoginView} />
       </div>
     );
